Split videos decorator for better encapsulation
