+++ app/controllers/videos_controller.rb
@@ -12,7 +12,7 @@ private
 
   # Use callbacks to share common setup or constraints between actions.
   def set_video
-    @video = VideoDecorator.new(Video.find(params[:id]))
+    @video = VideoShowDecorator.new(Video.find(params[:id]))
   end
 
 end
