+++ app/views/library/_video.html.erb
@@ -8,7 +8,7 @@
         <%= link_to video.title, video_path(video.id) %>
       </span>
       <% if video.needs_subscription? %>
-        <span class="rounded-full bg-indigo-500 uppercase px-2 py-1 text-xs font-bold flex items-center text-indigo-100">
+        <span class="rounded-full bg-indigo-500 uppercase px-2 py-1 text-xs font-bold flex items-center text-indigo-100 mt-1">
           subscrirption
         </span>
       <% end %>
